refactor: move obsidian-related ui components to obsidian module
