+++ src/obsidian/ui/editor-extension.ts
