+++ src/obsidian/ui/reminder.ts
@@ -2,9 +2,9 @@ import type { ReadOnlyReference } from 'model/ref';
 import type { DateTime } from 'model/time';
 import { App, Modal } from 'obsidian';
 import { SETTINGS } from 'obsidian/settings';
-import type { Reminder } from '../model/reminder';
-import type { Later } from '../model/time';
-import ReminderView from './components/Reminder.svelte';
+import type { Reminder } from '../../model/reminder';
+import type { Later } from '../../model/time';
+import ReminderView from 'ui/Reminder.svelte';
 const electron = require('electron');
 
 export class ReminderModal {
