+++ src/ui/DateTimeChooser.svelte
@@ -1,6 +1,6 @@
 <script lang="typescript">
-    import type { Reminders } from "model/reminder";
-    import { DateTime } from "model/time";
+    import type { Reminders } from "../model/reminder";
+    import { DateTime } from "../model/time";
     import moment from "moment";
     import type { Component } from "obsidian";
     import CalendarView from "./Calendar.svelte";
