+++ src/ui/timed-input-handler.ts
